# docs/api-by-nest/forum.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/post/{postId}": {
      "get": {
        "summary": "帖子详情",
        "deprecated": false,
        "description": "",
        "tags": ["post"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "pId": {
                          "type": "string"
                        },
                        "pViewCount": {
                          "type": "string"
                        },
                        "pCollectCount": {
                          "type": "string"
                        },
                        "pShareCount": {
                          "type": "string"
                        },
                        "pCommentCount": {
                          "type": "string"
                        },
                        "pContent": {
                          "type": "string"
                        },
                        "pImages": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "publishTime": {
                          "type": "string"
                        },
                        "isPublic": {
                          "type": "string"
                        },
                        "status": {
                          "type": "integer"
                        },
                        "username": {
                          "type": "string"
                        },
                        "userAvatar": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "pId",
                        "pViewCount",
                        "pCollectCount",
                        "pShareCount",
                        "pCommentCount",
                        "pContent",
                        "pImages",
                        "publishTime",
                        "isPublic",
                        "status",
                        "username",
                        "userAvatar"
                      ]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "pId": "p00076",
                    "pViewCount": "30",
                    "pCollectCount": "0",
                    "pShareCount": "1",
                    "pCommentCount": "0",
                    "pContent": "6666",
                    "pImages": [
                      "/uploads/2025-11-07_16-46-37_28cbbf3fac94510b.jpg",
                      "/uploads/2025-11-07_16-46-37_3415d3cf7c183e03.jpg",
                      "/uploads/2025-11-07_16-46-37_35e8676f15d78e7f.jpg",
                      "/uploads/2025-11-07_16-46-37_58458711eae3fa71.jpg",
                      "/uploads/2025-11-07_16-46-37_5687b7c61278a24a.jpg"
                    ],
                    "publishTime": "2025-11-07T08:46:37.000Z",
                    "isPublic": "true",
                    "status": 1,
                    "username": "默认名字",
                    "userAvatar": "/uploads/2025-12-20_19-07-36_38937f90635c9206.jpg"
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    }
                  },
                  "required": ["message", "error", "statusCode"]
                },
                "example": {
                  "message": "帖子不可见或已删除",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/user/{userId}": {
      "get": {
        "summary": "查询用户信息",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "example": "self"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "userAvatar": {
                          "type": "string"
                        },
                        "userEmail": {
                          "type": "string"
                        },
                        "registrationTime": {
                          "type": "string"
                        },
                        "follows": {
                          "type": "string"
                        },
                        "fans": {
                          "type": "string"
                        },
                        "sex": {
                          "type": "string"
                        },
                        "signature": {
                          "type": "string"
                        },
                        "backgroundImg": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "userId",
                        "username",
                        "userAvatar",
                        "userEmail",
                        "registrationTime",
                        "follows",
                        "fans",
                        "sex",
                        "signature",
                        "backgroundImg"
                      ]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "userId": "u00000",
                    "username": "asdf",
                    "userAvatar": "/uploads/2025-12-20_19-07-36_38937f90635c9206.jpg",
                    "userEmail": "chcaya@qq.com",
                    "registrationTime": "2025-07-27T06:08:36.000Z",
                    "follows": "4",
                    "fans": "2",
                    "sex": "boy",
                    "signature": "hhhhhhhhhhhhh",
                    "backgroundImg": "/uploads/2025-12-20_17-09-40_31e5daf42f2fa159.jpg"
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    }
                  },
                  "required": ["message", "error", "statusCode"]
                },
                "examples": {
                  "2": {
                    "summary": "异常示例",
                    "value": {
                      "message": "未找到该用户",
                      "error": "Not Found",
                      "statusCode": 404
                    }
                  },
                  "3": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "success",
                      "data": {
                        "userId": "u00000",
                        "username": "asdf",
                        "userAvatar": "/uploads/2025-12-20_19-07-36_38937f90635c9206.jpg",
                        "userEmail": "chcaya@qq.com",
                        "userPassword": "$2b$10$skMZ0PMvfdyzIkmX4FsZ1eWrsn6CXvPOFSkbN.ezy9nygfWc4XfEC",
                        "registrationTime": "2025-07-27T06:08:36.000Z",
                        "follows": "4",
                        "fans": "2",
                        "sex": "boy",
                        "signature": "hhhhhhhhhhhhh",
                        "backgroundImg": "/uploads/2025-12-20_17-09-40_31e5daf42f2fa159.jpg"
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/post/{userId}": {
      "get": {
        "summary": "查询用户帖子",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "example": "10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "pId": {
                                "type": "string"
                              },
                              "userId": {
                                "type": "string"
                              },
                              "pViewCount": {
                                "type": "string"
                              },
                              "pCollectCount": {
                                "type": "string"
                              },
                              "pShareCount": {
                                "type": "string"
                              },
                              "pCommentCount": {
                                "type": "string"
                              },
                              "pContent": {
                                "type": "string"
                              },
                              "pImages": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "publishTime": {
                                "type": "string"
                              },
                              "isPublic": {
                                "type": "string"
                              },
                              "status": {
                                "type": "integer"
                              },
                              "username": {
                                "type": "string"
                              },
                              "userAvatar": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["list", "total"]
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "list": [
                      {
                        "pId": "p00076",
                        "userId": "u00000",
                        "pViewCount": "30",
                        "pCollectCount": "0",
                        "pShareCount": "1",
                        "pCommentCount": "0",
                        "pContent": "6666",
                        "pImages": [
                          "/uploads/2025-11-07_16-46-37_28cbbf3fac94510b.jpg",
                          "/uploads/2025-11-07_16-46-37_3415d3cf7c183e03.jpg",
                          "/uploads/2025-11-07_16-46-37_35e8676f15d78e7f.jpg",
                          "/uploads/2025-11-07_16-46-37_58458711eae3fa71.jpg",
                          "/uploads/2025-11-07_16-46-37_5687b7c61278a24a.jpg"
                        ],
                        "publishTime": "2025-11-07T08:46:37.000Z",
                        "isPublic": "true",
                        "status": 1,
                        "username": "默认名字",
                        "userAvatar": "/uploads/2025-12-20_19-07-36_38937f90635c9206.jpg"
                      }
                    ],
                    "total": 1
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/user/collect/ids": {
      "get": {
        "summary": "查询用户收藏帖子的id数组",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": ["p00023"]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/search": {
      "get": {
        "summary": "搜索用户",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "userAvatar": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "userId": "u00003",
                      "username": "纳尼",
                      "userAvatar": "/uploads/2025-11-09_21-33-50_87eea2f7f7f1332f.jpg"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/friend": {
      "get": {
        "summary": "查询用户好友",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "userId": "u00000",
                      "followId": "u00001",
                      "username": "超级飞侠",
                      "userAvatar": "/uploads/default/default_avatar.jpg"
                    },
                    {
                      "userId": "u00000",
                      "followId": "u00002",
                      "username": "高雅人士",
                      "userAvatar": "/uploads/2025-11-09_14-42-42_dacb43ae3c952188.jpg"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/follows/{userId}": {
      "get": {
        "summary": "查询用户关注",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "followId": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "userAvatar": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "userId",
                          "followId",
                          "username",
                          "userAvatar"
                        ]
                      }
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "userId": "u00000",
                      "followId": "u00001",
                      "username": "超级飞侠",
                      "userAvatar": "/uploads/default/default_avatar.jpg"
                    },
                    {
                      "userId": "u00000",
                      "followId": "u00002",
                      "username": "高雅人士",
                      "userAvatar": "/uploads/2025-11-09_14-42-42_dacb43ae3c952188.jpg"
                    },
                    {
                      "userId": "u00000",
                      "followId": "u00003",
                      "username": "纳尼",
                      "userAvatar": "/uploads/2025-11-09_21-33-50_87eea2f7f7f1332f.jpg"
                    },
                    {
                      "userId": "u00000",
                      "followId": "u00007",
                      "username": "默认名字",
                      "userAvatar": "/uploads/default/default_avatar.jpg"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/fans/{userId}": {
      "get": {
        "summary": "查询用户粉丝",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "followId": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "userAvatar": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "userId",
                          "followId",
                          "username",
                          "userAvatar"
                        ]
                      }
                    }
                  },
                  "required": ["code", "message", "data"]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "userId": "u00001",
                      "followId": "u00000",
                      "username": "超级飞侠",
                      "userAvatar": "/uploads/default/default_avatar.jpg"
                    },
                    {
                      "userId": "u00002",
                      "followId": "u00000",
                      "username": "高雅人士",
                      "userAvatar": "/uploads/2025-11-09_14-42-42_dacb43ae3c952188.jpg"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/user/register": {
      "post": {
        "summary": "注册",
        "deprecated": false,
        "description": "",
        "tags": ["user"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                },
                "example": {
                  "code": 0,
                  "message": "success"
                }
              }
            },
            "headers": {}
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    }
                  },
                  "required": ["message", "error", "statusCode"]
                },
                "example": {
                  "message": "该邮箱已被注册",
                  "error": "Conflict",
                  "statusCode": 409
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}